Moved payload_generator_helper.py, added SLEEP to meterpreter python
