+++ Tools/Evasion/payloads/python/payload_generator_helper.py
