+++ Tools/Evasion/payloads/python/shellcode_inject/letter_substitution.py
@@ -17,7 +17,7 @@ import random
 import string
 from datetime import date
 from datetime import timedelta
-from Tools.Evasion.payloads.python.shellcode_inject import payload_generator_helper
+from Tools.Evasion.payloads.python import payload_generator_helper
 from Tools.Evasion.evasion_common import encryption
 from Tools.Evasion.evasion_common import evasion_helpers
 from Tools.Evasion.evasion_common import shellcode_help
